Set UTC as the default timezone
Since the all of the return value of `Time.now.zone` are not always recognized
by TZInfo, `Time.find_zone!` should be avoided. The reason why it does not
recognize is that `Time.now.zone` returns the timezone string generated by
`strftime` while TZInfo uses zoneinfo library, and there is no consistency
between them. Moreover, it is problematic that there is no conformance with
`ActiveRecord::Base.default_timezone` to `Time.now.zone`.

This commit resorts to set the default timezone to UTC and let users specify
their timezone in the configuration files as documented in README.