Do not split the rsync server and the remote path
It is troublesome to split the rsync server and the remote path under the
situation that the system uses rsyncd.

It also moves rsync settings to the workers and clients from the server. This is
because It is possible that the rsync command in the workers and the one in the
clients are different.