Do average calculation in another thread
- Extracted average calulation to StatisticsUpdater.
- Added StatisticsUpdaterQueue as a queue for ditto

Updating averages is too slow; yes, I know we should fix that...

The problem is, `rrrspec waitfor` command waits the calculation. Not
only a taskset completion. Because it waits persistence.

Persistence phase may late because persisting run after previous average
calculation. This makes a lag. So `rrrspec waitfor` waits for
the previous calulation too.

This extract statistics calculation to the another thread, and allow
Persister to work persisting of another completed taskset faster.